refactor: SetHostInfo component refactoring
